·修改版本号为0.112 ·修改日期选择方式 ·完成日程显示功能 ·修改millitime为Long

+++ app/src/main/java/com/wxy3265/vmeknowledge/Knowledge.kt
@@ -1,5 +1,5 @@
 package com.wxy3265.vmeknowledge
 
 class Knowledge(val Content: String, val CreateDate: String,val ReviewDate:String,val Id: Int,
-                val StudyTimes: Int, val MilliTime: Int, val tag:String) {
+                val StudyTimes: Int, val MilliTime: Long, val tag:String) {
 }
