+++ app/src/main/java/com/wxy3265/vmeknowledge/Schedule.kt
@@ -1,7 +1,7 @@
 package com.wxy3265.vmeknowledge
 
 class Schedule (val content: String, val id: Int, val startTime: String, val endTime: String,
-                val startMilliTime: Int, val endMilliTime:Int, val s_notification: String, val tag: String) {
+                val startMilliTime: Long, val endMilliTime:Long, val s_notification: String, val tag: String) {
 
 }
