+++ app/src/main/java/com/wxy3265/vmeknowledge/StudyTagActivity.kt
@@ -37,7 +37,7 @@ class StudyTagActivity : AppCompatActivity() {
                 val reviewdate = cursor.getString(cursor.getColumnIndex("reviewdate"))
                 val id = cursor.getInt(cursor.getColumnIndex("id"))
                 val studyTimes = cursor.getInt(cursor.getColumnIndex("studytimes"))
-                val milliTime = cursor.getInt(cursor.getColumnIndex("milliTime"))
+                val milliTime = cursor.getLong(cursor.getColumnIndex("milliTime"))
                 val tag = cursor.getString(cursor.getColumnIndex("tag"))
                 Log.d(TAG, "onCreate: knowledge: " + content)
                 val ktag = Tag(tag)
